moved validation from Game class to Validation class
